+++ container-images/mds-agency/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @container-images/mds-agency
 
+## 0.3.10
+
+### Patch Changes
+
+- 8a3e63d0: Breaking interface changes due to ioredis version bump
+- Updated dependencies [8a3e63d0]
+  - @mds-core/mds-agency@0.5.23
+  - @mds-core/mds-webpack-config@0.1.14
+
 ## 0.3.9
 
 ### Patch Changes