+++ container-images/mds-collector-service/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @container-images/mds-collector-service
 
+## 0.2.9
+
+### Patch Changes
+
+- 8a3e63d0: Breaking interface changes due to ioredis version bump
+- Updated dependencies [8a3e63d0]
+  - @mds-core/mds-collector-service@0.4.36
+  - @mds-core/mds-webpack-config@0.1.14
+
 ## 0.2.8
 
 ### Patch Changes