+++ packages/mds-config-api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mds-core/mds-config-api
 
+## 0.4.14
+
+### Patch Changes
+
+- 8a3e63d0: Breaking interface changes due to ioredis version bump
+- Updated dependencies [8a3e63d0]
+  - @mds-core/mds-api-server@0.4.15
+  - @mds-core/mds-config-service@0.4.14
+  - @mds-core/mds-logger@0.5.8
+  - @mds-core/mds-service-helpers@0.6.11
+  - @mds-core/mds-utils@0.5.3
+
 ## 0.4.13
 
 ### Patch Changes