+++ packages/mds-geography-author-api/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @mds-core/mds-geography-author-api
 
+## 0.4.36
+
+### Patch Changes
+
+- 8a3e63d0: Breaking interface changes due to ioredis version bump
+- Updated dependencies [8a3e63d0]
+  - @mds-core/mds-api-helpers@0.3.2
+  - @mds-core/mds-api-server@0.4.15
+  - @mds-core/mds-geography-service@0.10.33
+  - @mds-core/mds-logger@0.5.8
+  - @mds-core/mds-schema-validators@0.5.15
+  - @mds-core/mds-service-helpers@0.6.11
+  - @mds-core/mds-utils@0.5.3
+  - @mds-core/mds-types@0.9.3
+
 ## 0.4.35
 
 ### Patch Changes